modern3: make compiled sequence-length domains LLP64-safe (M3-SOL-R1)
SOL HIGH finding D1: on Win64 (LLP64) sizeof(long)==sizeof(unsigned long)==4,
so every len(seq)/pattern-length scalar declared 'long'/'unsigned long'
wrapped reported lengths >= 2**32 (a 2**32 probe behaved as the empty
pattern).  Repair, per docs/modernization/NUMERIC_WIDTH_POLICY.md:

- BasicBWT.pyx: countOccurrencesOfSeq / findIndicesOfStr /
  findIndicesOfRegex / findStrWithError / findPatternWithError /
  findKmerWithError / findKmerWithErrors (+stackIndex) / findRangeOfStr /
  countPileup / countPileup_c / countSeqMatches / countStrandedSeqMatches /
  countStrandedSeqMatchesNoOther / findKmerThreshold(Stranded) /
  findKTOtherStranded / recoverString loop counter: len()-derived lengths
  and pattern-position counters now Py_ssize_t; the C helpers
  countOccurrencesOfSeq_c/findRangeOfStr_c take Py_ssize_t seqLen
  (BasicBWT.pxd updated).  Symbol/row/score domains stay narrow per policy.
- AlignmentUtil.pyx: oLen/mLen and matrix loop coordinates in fullAlign,
  fullED_score, fullED_minimize (incl. argmin*/scoreLen), fullAlign_noGO,
  alignChanges now Py_ssize_t; scoring constants remain narrow.
- RLE_BWTCython.pyx: pattern-position counters of the three LCP-assisted
  query paths now Py_ssize_t.
- MultiStringBWTCython.pyx: maxSeqLen/uniformSeqLen read-length sentinels
  now Py_ssize_t.

Adds sol_r1_seq_length_transport / sol_r1_align_length_transport compiled
probes and compat/tests/test_m3_sol_r1_signed_lengths.py: nonallocating
2^31-1..2^32+12345 boundary probes plus Byte/RLE/LCP-assisted production
query semantics vs independent naive oracles.

+++ packages/msbwt-modern3/MUSCython/BasicBWT.pxd
@@ -95,9 +95,9 @@ cdef class BasicBWT(object):
     cdef bwtRange getOccurrenceOfCharAtRange(BasicBWT self, np.uint8_t sym, bwtRange inRange) nogil
     cdef bwtRange findRangeOfStr(BasicBWT self, object seq)
     cdef np.ndarray[np.uint64_t, ndim=1, mode='c'] countPileup_c(BasicBWT self, object seq, Py_ssize_t kmerSize)
-    cdef np.uint64_t countOccurrencesOfSeq_c(BasicBWT self, unsigned char * seq_view, np.uint64_t seqLen, np.uint64_t mc=*)
+    cdef np.uint64_t countOccurrencesOfSeq_c(BasicBWT self, unsigned char * seq_view, Py_ssize_t seqLen, np.uint64_t mc=*)
     cdef np.uint64_t getOccurrenceOfCharAtIndex_c(BasicBWT self, np.uint8_t sym, np.uint64_t index)
-    cdef bwtRange findRangeOfStr_c(BasicBWT self, unsigned char * seq_view, np.uint64_t seqLen)
+    cdef bwtRange findRangeOfStr_c(BasicBWT self, unsigned char * seq_view, Py_ssize_t seqLen)
     
     #the following functions require an LCP array
     cpdef tuple countSeqMatches(BasicBWT self, object seq, Py_ssize_t kmerSize)
